+++ crates/swc_ecma_minifier/src/compress/pure/misc.rs
@@ -3,7 +3,10 @@ use std::{fmt::Write, num::FpCategory};
 use swc_atoms::js_word;
 use swc_common::{iter::IdentifyLast, util::take::Take, Span, DUMMY_SP};
 use swc_ecma_ast::*;
-use swc_ecma_utils::ExprExt;
+use swc_ecma_utils::{
+    ExprExt, Type,
+    Value::{self, Known},
+};
 
 use super::Pure;
 use crate::compress::{
@@ -963,6 +966,48 @@ impl Pure<'_> {
                 *e = l.take();
             }
             _ => {}
+        }
+    }
+
+    ///
+    /// - `a ? true : false` => `!!a`
+    pub(super) fn compress_useless_cond_expr(&mut self, expr: &mut Expr) {
+        let cond = match expr {
+            Expr::Cond(c) => c,
+            _ => return,
+        };
+
+        let lt = cond.cons.get_type();
+        let rt = cond.alt.get_type();
+        match (lt, rt) {
+            (Known(Type::Bool), Known(Type::Bool)) => {}
+            _ => return,
+        }
+
+        let lb = cond.cons.as_pure_bool(&self.expr_ctx);
+        let rb = cond.alt.as_pure_bool(&self.expr_ctx);
+
+        let lb = match lb {
+            Value::Known(v) => v,
+            Value::Unknown => return,
+        };
+        let rb = match rb {
+            Value::Known(v) => v,
+            Value::Unknown => return,
+        };
+
+        // `cond ? true : false` => !!cond
+        if lb && !rb {
+            self.negate(&mut cond.test, false, false);
+            self.negate(&mut cond.test, false, false);
+            *expr = *cond.test.take();
+            return;
+        }
+
+        // `cond ? false : true` => !cond
+        if !lb && rb {
+            self.negate(&mut cond.test, false, false);
+            *expr = *cond.test.take();
         }
     }
 }
