+++ crates/swc_ecma_minifier/src/compress/pure/mod.rs
@@ -370,6 +370,8 @@ impl VisitMut for Pure<'_> {
         self.optimize_nullish_coalescing(e);
 
         self.compress_negated_bin_eq(e);
+
+        self.compress_useless_cond_expr(e);
     }
 
     fn visit_mut_expr_stmt(&mut self, s: &mut ExprStmt) {
